Native: update isl download URL in a toolchain builder config
The previous URL no longer works. The isl project has moved hosting to
sourceforge.io. Update the mirror accordingly.
This is fixed in Crosstool-NG 1.25.0, but the URL is included in config
and needs to be updated there.

Note: to make the rest of toolchains buildable, the same change should
be applied to their configs.

See also https://github.com/crosstool-ng/crosstool-ng/pull/1614.